pyproject.toml
[tool.poetry]
name = "dw-natixis"
version = "0.1.0"
description = "API for uploading, storing, embedding and querying files."
authors = ["Denis <dchen@ampliforce.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
fastapi = "^0.111.0"
minio = "^7.2.7"
python-dotenv = "^1.0.1"
pydantic = "^2.7.1"
aiohttp = "^3.9.5"
aiofiles = "^23.2.1"
tiktoken = "^0.7.0"
langchain = "^0.1.20"
langchain-pinecone = "^0.1.0"
langchain-openai = "^0.1.6"
openai = "^1.30.1"
async-lru = "^2.0.4"
requests = "^2.32.1"
pypdf = "^4.2.0"
boto3 = "^1.34.118"
semantic-chunkers = "^0.0.6"
cohere = "^5.5.6"
langchain-cohere = "^0.1.5"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
flake8 = "^7.0.0"
mypy = "^1.10.0"
isort = "^5.13.2"
ipykernel = "^6.29.4"

[tool.isort]
line_length = 120
profile = "black"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
